# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Better Open Snow API",
    "description": "Snow + weather data for Epic/Ikon resorts (v0 MVP)",
    "version": "0.0.0"
  },
  "paths": {
    "/resorts": {
      "get": {
        "tags": [
          "resorts"
        ],
        "summary": "List Resorts",
        "operationId": "list_resorts_resorts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResortList"
                }
              }
            }
          }
        }
      }
    },
    "/resorts/{resort_id}": {
      "get": {
        "tags": [
          "resorts"
        ],
        "summary": "Get Resort",
        "operationId": "get_resort_resorts__resort_id__get",
        "parameters": [
          {
            "name": "resort_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resort Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResortDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resorts/{resort_id}/forecast": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "Get Forecast",
        "operationId": "get_forecast_resorts__resort_id__forecast_get",
        "parameters": [
          {
            "name": "resort_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resort Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forecast"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resorts/{resort_id}/observations": {
      "get": {
        "tags": [
          "observations"
        ],
        "summary": "Get Observations",
        "operationId": "get_observations_resorts__resort_id__observations_get",
        "parameters": [
          {
            "name": "resort_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resort Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Observations"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/debug/firestore-ping": {
      "get": {
        "summary": "Firestore Ping",
        "operationId": "firestore_ping_debug_firestore_ping_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Firestore Ping Debug Firestore Ping Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DailyPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "snow_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snow In"
          },
          "temp_hi_f": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp Hi F"
          },
          "temp_lo_f": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp Lo F"
          },
          "wind_mph_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Mph Max"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "DailyPoint"
      },
      "Forecast": {
        "properties": {
          "resort_id": {
            "type": "string",
            "title": "Resort Id"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "hourly": {
            "items": {
              "$ref": "#/components/schemas/HourlyPoint"
            },
            "type": "array",
            "title": "Hourly"
          },
          "daily": {
            "items": {
              "$ref": "#/components/schemas/DailyPoint"
            },
            "type": "array",
            "title": "Daily"
          },
          "sources": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Sources"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "resort_id",
          "generated_at",
          "hourly",
          "daily",
          "sources"
        ],
        "title": "Forecast"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HourlyPoint": {
        "properties": {
          "t": {
            "type": "string",
            "title": "T"
          },
          "temp_f": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp F"
          },
          "wind_mph": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Mph"
          },
          "snowfall_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snowfall In"
          },
          "precip_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precip In"
          },
          "visibility_mi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility Mi"
          }
        },
        "type": "object",
        "required": [
          "t"
        ],
        "title": "HourlyPoint"
      },
      "Observations": {
        "properties": {
          "resort_id": {
            "type": "string",
            "title": "Resort Id"
          },
          "snotel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnotelObservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "resort_id"
        ],
        "title": "Observations"
      },
      "ResortBase": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "elevation_ft": {
            "type": "integer",
            "title": "Elevation Ft"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon",
          "elevation_ft"
        ],
        "title": "ResortBase"
      },
      "ResortDetail": {
        "properties": {
          "resort_id": {
            "type": "string",
            "title": "Resort Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "pass": {
            "type": "string",
            "enum": [
              "epic",
              "ikon"
            ],
            "title": "Pass"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "base": {
            "$ref": "#/components/schemas/ResortBase"
          },
          "summit": {
            "$ref": "#/components/schemas/ResortElevation"
          },
          "snotel_triplet": {
            "type": "string",
            "title": "Snotel Triplet"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          }
        },
        "type": "object",
        "required": [
          "resort_id",
          "name",
          "pass",
          "state",
          "base",
          "summit",
          "snotel_triplet",
          "timezone"
        ],
        "title": "ResortDetail",
        "description": "Full metadata for the detail screen."
      },
      "ResortElevation": {
        "properties": {
          "elevation_ft": {
            "type": "integer",
            "title": "Elevation Ft"
          }
        },
        "type": "object",
        "required": [
          "elevation_ft"
        ],
        "title": "ResortElevation"
      },
      "ResortList": {
        "properties": {
          "resorts": {
            "items": {
              "$ref": "#/components/schemas/ResortSummary"
            },
            "type": "array",
            "title": "Resorts"
          }
        },
        "type": "object",
        "required": [
          "resorts"
        ],
        "title": "ResortList"
      },
      "ResortSummary": {
        "properties": {
          "resort_id": {
            "type": "string",
            "title": "Resort Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "pass": {
            "type": "string",
            "enum": [
              "epic",
              "ikon"
            ],
            "title": "Pass"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "base_elevation_ft": {
            "type": "integer",
            "title": "Base Elevation Ft"
          },
          "summit_elevation_ft": {
            "type": "integer",
            "title": "Summit Elevation Ft"
          },
          "latest_snow_24h_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Snow 24H In"
          },
          "snow_depth_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snow Depth In"
          }
        },
        "type": "object",
        "required": [
          "resort_id",
          "name",
          "pass",
          "state",
          "base_elevation_ft",
          "summit_elevation_ft"
        ],
        "title": "ResortSummary",
        "description": "Compact row for the list screen."
      },
      "SnotelObservation": {
        "properties": {
          "station_triplet": {
            "type": "string",
            "title": "Station Triplet"
          },
          "observed_at": {
            "type": "string",
            "title": "Observed At"
          },
          "swe_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Swe In"
          },
          "snow_depth_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snow Depth In"
          },
          "snow_24h_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snow 24H In"
          },
          "air_temp_f": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Air Temp F"
          }
        },
        "type": "object",
        "required": [
          "station_triplet",
          "observed_at"
        ],
        "title": "SnotelObservation"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}